Fix init file paths
DESTDIR is used by distribution packaging to determine paths for make to
install too, which always is some other location than the root of the
filesystem. The resulting package will then install it to the root on
user systems. Thus, using DESTDIR in the Makefile causes wrong paths to
be used

Also, the binaries are always installed to $PREFIX/bin, not
$PREFIX/linux-<arch>-<whatever>/bin, and same with all other files
(config, apps, etc). Thus remove $(TARGET_BIN) from the bin path